:recycle: [Refactor] toggleBookmark 메서드 위치 변경, dto 패키지 구조 변경에 따른 import 문제 해결
