+++ src/main/java/kcs/funding/fundingboost/domain/controller/ItemController.java
@@ -2,7 +2,6 @@ package kcs.funding.fundingboost.domain.controller;
 
 
 import java.util.List;
-import kcs.funding.fundingboost.domain.dto.common.CommonSuccessDto;
 import kcs.funding.fundingboost.domain.dto.global.ResponseDto;
 import kcs.funding.fundingboost.domain.dto.response.shopping.ShopDto;
 import kcs.funding.fundingboost.domain.dto.response.shoppingDetail.ItemDetailDto;
@@ -11,7 +10,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -39,14 +37,5 @@ public class ItemController {
     public ResponseDto<ItemDetailDto> showItemDetail(@RequestParam(name = "memberId") Long memberId,
                                                      @PathVariable(name = "itemId") Long itemId) {
         return ResponseDto.ok(itemService.getItemDetail(memberId, itemId));
-    }
-
-    /**
-     * 상품 좋아요
-     */
-    @PostMapping("/like/{itemId}")
-    public ResponseDto<CommonSuccessDto> itemLike(@RequestParam(name = "memberId") Long memberId,
-                                                  @PathVariable(name = "itemId") Long itemId) {
-        return ResponseDto.ok(itemService.toggleItemLike(memberId, itemId));
     }
 }
