+++ src/main/java/kcs/funding/fundingboost/domain/service/ItemService.java
@@ -1,20 +1,16 @@
 package kcs.funding.fundingboost.domain.service;
 
 import static kcs.funding.fundingboost.domain.exception.ErrorCode.NOT_FOUND_ITEM;
-import static kcs.funding.fundingboost.domain.exception.ErrorCode.NOT_FOUND_MEMBER;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import kcs.funding.fundingboost.domain.dto.common.CommonSuccessDto;
 import kcs.funding.fundingboost.domain.dto.response.shopping.ShopDto;
 import kcs.funding.fundingboost.domain.dto.response.shoppingDetail.ItemDetailDto;
 import kcs.funding.fundingboost.domain.entity.Bookmark;
 import kcs.funding.fundingboost.domain.entity.Item;
-import kcs.funding.fundingboost.domain.entity.Member;
 import kcs.funding.fundingboost.domain.exception.CommonException;
 import kcs.funding.fundingboost.domain.repository.ItemRepository;
-import kcs.funding.fundingboost.domain.repository.MemberRepository;
 import kcs.funding.fundingboost.domain.repository.bookmark.BookmarkRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -26,7 +22,6 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j
 @RequiredArgsConstructor
 public class ItemService {
-    private final MemberRepository memberRepository;
     private final ItemRepository itemRepository;
     private final BookmarkRepository bookmarkRepository;
 
@@ -49,24 +44,5 @@ public class ItemService {
 
             return ItemDetailDto.fromEntity(item, false);
         }
-    }
-
-    @Transactional
-    public CommonSuccessDto toggleItemLike(Long memberId, Long itemId) {
-        Optional<Bookmark> optionalBookmark = bookmarkRepository.findBookmarkByMemberAndItem(memberId, itemId);
-
-        if (optionalBookmark.isPresent()) {
-            bookmarkRepository.delete(optionalBookmark.get());
-        } else {
-            Member member = memberRepository.findById(memberId)
-                    .orElseThrow(() -> new CommonException(NOT_FOUND_MEMBER));
-
-            Item item = itemRepository.findById(itemId)
-                    .orElseThrow(() -> new CommonException(NOT_FOUND_ITEM));
-
-            Bookmark bookmark = Bookmark.createBookmark(member, item);
-            bookmarkRepository.save(bookmark);
-        }
-        return CommonSuccessDto.fromEntity(true);
     }
 }
